requirements.txt
pandas
requests
snowflake-connector-python
python-dotenv